+++ pelutils/types.py
@@ -6,6 +6,9 @@ import numpy.typing as npt
 # Common numpy types
 AnyArray = npt.NDArray[Any]  # pyright: ignore[reportExplicitAny]
 BoolArray = npt.NDArray[np.bool_]
+BytesArray = npt.NDArray[np.bytes_]
 ComplexArray = npt.NDArray[np.complexfloating[Any]]  # pyright: ignore[reportExplicitAny, reportInvalidTypeArguments]
 FloatArray = npt.NDArray[np.floating[Any]]  # pyright: ignore[reportExplicitAny]
 IntArray = npt.NDArray[np.integer[Any]]  # pyright: ignore[reportExplicitAny]
+ObjectArray = npt.NDArray[np.object_]
+StringArray = npt.NDArray[np.str_]